Use SIGVTALRM to kill timeouting processes (default: use SIGKILL)

+++ subproc.c
@@ -327,10 +327,24 @@ void subproc_checkTimeLimit(honggfuzz_t * hfuzz, fuzzer_t * fuzzer)
 
     int64_t curMillis = util_timeNowMillis();
     int64_t diffMillis = curMillis - fuzzer->timeStartedMillis;
-    if (diffMillis > (hfuzz->tmOut * 1000)) {
-        LOG_W("PID %d took too much time (limit %ld s). Sending SIGKILL",
-              fuzzer->pid, hfuzz->tmOut);
+
+    if (fuzzer->tmOutSignaled && (diffMillis > ((hfuzz->tmOut + 1) * 1000))) {
+        /* Has this instance been already signaled due to timeout? Just, SIGKILL it */
+        LOG_W("PID %d has already been signaled due to timeout. Killing it with SIGKILL",
+              fuzzer->pid);
         kill(fuzzer->pid, SIGKILL);
+        return;
+    }
+
+    if ((diffMillis > (hfuzz->tmOut * 1000)) && fuzzer->tmOutSignaled == false) {
+        fuzzer->tmOutSignaled = true;
+        LOG_W("PID %d took too much time (limit %ld s). Killing it with %s", fuzzer->pid,
+              hfuzz->tmOut, hfuzz->tmout_vtalrm ? "SIGVTALRM" : "SIGKILL");
+        if (hfuzz->tmout_vtalrm) {
+            kill(fuzzer->pid, SIGVTALRM);
+        } else {
+            kill(fuzzer->pid, SIGKILL);
+        }
         ATOMIC_POST_INC(hfuzz->timeoutedCnt);
     }
 }
